Remove maintenance badge on Rutie
Rutie is actively maintained but the badge of maintenance status will go stale every year that rolls around.  Either I the maintainer would need to update this here every year, or just do this PR to remove it.  Seeing as how I have no guarantee that future PRs would continue to be merged here in the years to come I think it's wiser if I simply remove it now.